added G loading infobox, removed extraneous TaskStatistics calculation, fixed task distance to go when AAT enabled, settings dialog layout improvements

+++ Common/Source/Utils.cpp
@@ -329,7 +329,7 @@ void ReadRegistrySettings(void)
 
   Temp = 500;
   GetFromRegistry(szRegistryNettoSpeed,&Temp);
-  NettoSpeed = (double)Temp;
+  NettoSpeed = Temp;
 
 }
